updating all mvps

+++ server/app/controllers/parcels_controller.rb
@@ -97,45 +97,26 @@ class ParcelsController < ApplicationController
   
 
   def destroy
-    @parcel = Parcel.find_by(id: params[:id])
+    @parcel = current_user.parcels.find_by(id: params[:id])
 
     if @parcel.nil?
       render json: { error: "Parcel not found" }, status: :not_found
       return
     end
 
-    if @parcel.destroy
-      render json: { message: 'Parcel deleted successfully' }, status: :ok
+    if @parcel.status == 'pending' # Check parcel status before deletion
+      if @parcel.destroy
+        render json: { message: 'Parcel deleted successfully' }, status: :ok
+      else
+        render json: { error: @parcel.errors.full_messages.join(', ') }, status: :unprocessable_entity
+      end
     else
-      render json: { error: @parcel.errors.full_messages.join(', ') }, status: :unprocessable_entity
+      render json: { error: "Parcel cannot be deleted as it's not in pending status" }, status: :unprocessable_entity
     end
   end
 
-  # def update_status_and_location
-  #   @parcel = Parcel.find_by(id: params[:id])
 
-  #   if @parcel.nil?
-  #     render json: { error: "Parcel not found" }, status: :not_found
-  #     return
-  #   end
-
-  #   old_status = @parcel.status
-  #   new_status = params[:status]
-
-  #   if @parcel.update(status: new_status, location: params[:location])
-  #     # Send SMS notification if the status is updated by admin
-  #     if current_user&.is_admin && old_status != new_status
-  #       sender_phone_number = @parcel.receiver_phone_number
-  #       message = "Your parcel status has been updated to #{new_status}."
-  #       send_sms(sender_phone_number, message)  # Use the SMSSender module's method
-  #     end
-
-  #     render json: { message: 'Parcel status and location updated successfully' }, status: :ok
-  #   else
-  #     render json: { error: @parcel.errors.full_messages.join(', ') }, status: :unprocessable_entity
-  #   end
-  # end
-
+ 
 
   def update_status_and_location
     @parcel = Parcel.find_by(id: params[:id])
